feature/auth

+++ service/contacts.js
@@ -1,30 +1,45 @@
 const Contact = require("../models/contact");
 
-const getContacts = () => {
-  return Contact.find();
-};
+const getContacts = (owner, skip, limit, favorite) => {
+  const filters = { owner };
+  if (favorite !== undefined) {
+    filters.favorite = favorite;
+  }
 
-const getContactById = (contactId) => {
-  return Contact.findOne({ _id: contactId });
-};
-
-const removeContact = (contactId) => {
-  return Contact.findOneAndRemove({ _id: contactId });
-};
-
-const createContact = (fields) => {
-  return Contact.create(fields);
-};
-
-const replaceContact = (contactId, fields) => {
-  return Contact.findOneAndReplace({ _id: contactId }, fields, {
-    new: true,
-    runValidators: true,
+  return Contact.find(filters, "_id name email phone favorite", {
+    skip,
+    limit,
   });
 };
 
-const updateContact = (contactId, fields) => {
-  return Contact.findOneAndUpdate({ _id: contactId }, fields, {
+const getContactById = (owner, contactId) => {
+  return Contact.findOne(
+    { owner, _id: contactId },
+    "_id name email phone favorite"
+  );
+};
+
+const createContact = (owner, fields) => {
+  return Contact.create({ ...fields, owner });
+};
+
+const removeContact = (owner, contactId) => {
+  return Contact.findOneAndRemove({ owner, _id: contactId });
+};
+
+const replaceContact = (owner, contactId, fields) => {
+  return Contact.findOneAndReplace(
+    { owner, _id: contactId },
+    { ...fields, owner },
+    {
+      new: true,
+      runValidators: true,
+    }
+  );
+};
+
+const updateContact = (owner, contactId, fields) => {
+  return Contact.findOneAndUpdate({ owner, _id: contactId }, fields, {
     new: true,
     unValidators: true,
   });
